projects/quickshop-mvp-test4/quickshop-mvp/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic[email]==2.5.0
pydantic-settings==2.1.0
redis==5.0.1
stripe==7.8.0
sendgrid==6.10.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
python-dotenv==1.0.0
email-validator==2.1.0
pillow==10.1.0
boto3==1.34.0
sentry-sdk[fastapi]==1.38.0